Fixed narrowing warning

+++ BattlePong/src/States/StateBase.cpp
@@ -165,7 +165,7 @@ void StateBase::reset(std::shared_ptr<PongFrame> frame)
 
     float paddleOffset = Settings::wallThickness * 4;
     frame->ball->setPosition({static_cast<float>(window.getView().getSize().x) / 2, static_cast<float>(window.getView().getSize().y) / 2});
-    frame->ball->setVelocity({x, y});
+    frame->ball->setVelocity({static_cast<float>(x), static_cast<float>(y)});
     frame->leftPaddle->setPosition(sf::Vector2f(paddleOffset - Settings::wallThickness, static_cast<float>(window.getView().getSize().y) / 2));
     frame->rightPaddle->setPosition(sf::Vector2f(static_cast<float>(window.getView().getSize().x)  - paddleOffset, static_cast<float>(window.getView().getSize().y) / 2));
     frame->rightPaddle->handleInput(*frame->ball);
@@ -216,7 +216,7 @@ void StateBase::Init(std::shared_ptr<GU::Engin::Frame> frame)
 void StateBase::sfEvent(GU::Engin::Engin& engin, const sf::Event &event, std::shared_ptr<GU::Engin::Frame> frame)
 {
   UNUSED(frame);
-  
+
     switch (event.type)
     {
